Add main nav reducer
