--- templates/src/redux/actions/nav.js
@@ -1,8 +1,0 @@
-import keys from '../keys';
-
-export function setIsMobileMenuOpen(isOpen) {
-  return {
-    type: keys.SET_IS_MOBILE_MENU_OPEN,
-    isOpen
-  };
-}
