+++ templates/src/redux/keys.js
@@ -4,7 +4,7 @@ export default {
   SET_IS_MOBILE_LAYOUT: 'SET_IS_MOBILE_LAYOUT',
   SET_IS_PHONE_LAYOUT: 'SET_IS_PHONE_LAYOUT',
   SET_IS_TABLET_LAYOUT: 'SET_IS_TABLET_LAYOUT',
-  LANDING_LOADED: 'LANDING_LOADED',
   SET_IS_MOBILE_MENU_OPEN: 'SET_IS_MOBILE_MENU_OPEN',
+  LANDING_LOADED: 'LANDING_LOADED',
   TEST_ASYNC: 'TEST_ASYNC'
 };
